// packages/contracts/src/native.generated.ts
export const NATIVE_TOKEN_PROFILES = [
  {
    networkId: "ethereum-mainnet",
    tokenAddress: "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48",
    decimals: 6,
    addressCodeSha256: "c1bcd4dc4f30c4dcf6c9fe462abeaf2efefc065dcb4dec74996e471f1ce219b4",
    implementationAddress: "0x43506849d7c04f9138d1a2050bbf3a0c054402dd",
    implementationSlot: "0x7050c9e0f4ca769c69bd3a8ef740bc37934f8e2c036e5a723fd8ee048ed3f8c3",
    implementationCodeSha256: "21a831a7cff1a353b295490f43a66e48e55d1c76c60cd89b24138efd8e8b7c8a",
    extraReadAbi: [],
  },
  {
    networkId: "ethereum-mainnet",
    tokenAddress: "0x18084fba666a33d37592fa2633fd49a74dd93a88",
    decimals: 18,
    addressCodeSha256: "586b8f5bb9ce01b4ec70bc0adb1ffd3480468f8ffcebcc716bd05942fb8724c0",
    implementationAddress: null,
    implementationSlot: null,
    implementationCodeSha256: null,
    extraReadAbi: [],
  },
  {
    networkId: "mezo-mainnet",
    tokenAddress: "0x7b7c000000000000000000000000000000000000",
    decimals: 18,
    addressCodeSha256: "c83521842d34e82c2bb74f3dd3a61c98e75e337e299aa80b6fa18d64652b1826",
    implementationAddress: null,
    implementationSlot: null,
    implementationCodeSha256: null,
    extraReadAbi: [],
  },
  {
    networkId: "mezo-mainnet",
    tokenAddress: "0x04671c72aab5ac02a03c1098314b1bb6b560c197",
    decimals: 6,
    addressCodeSha256: "4bc884c60f64f3486dd8c1a31ede2542f7417edc9e94730d9a86ae386cb82f65",
    implementationAddress: "0x09d42a1c4792546c26451af2d373bf0d4c545baf",
    implementationSlot: "0x360894a13ba1a3210667c828492db98dca3e2076cc3735a920a3ca505d382bbc",
    implementationCodeSha256: "378d40bcb07169de058746028e58f77128dcde6ff1660a70dbe338a7e700abc2",
    extraReadAbi: [
      {
        inputs: [],
        name: "minter",
        outputs: [{ internalType: "address", name: "", type: "address" }],
        stateMutability: "view",
        type: "function",
      },
    ],
  },
] as const;